Added support for the 'sendMode' feature added in documentation version 1.4 but not configurable in the PHP library.
sendMode is a great (and probably the only) way at the moment to simulate a 'sandbox' to message sending in message media.
It benefits the developer to be able to controll this flag to simulate sending SMS. However, there's no available way in this library
To use this feature apart from hacking the MMSoap.php which is not ideal. Therefore I would like to make this variable as a parameter
in sendMessages.

Also helped with some formatting into psr-2 standards.